The inventory is now working

+++ src/rosecommon/include/packets/cli_equipitem.h
@@ -11,19 +11,19 @@ class CliEquipItem : public CRosePacket {
 	public:
 		CliEquipItem();
 		CliEquipItem(uint8_t buffer[MAX_PACKET_SIZE]);
-		CliEquipItem(uint8_t slotTo, uint8_t slotFrom);
+		CliEquipItem(int16_t slotTo, int16_t slotFrom);
 
 		virtual ~CliEquipItem() = default;
 
-		uint8_t slotTo() const;
-		uint8_t slotFrom() const;
+		int16_t slotTo() const;
+		int16_t slotFrom() const;
 
 	protected:
 		virtual void pack() override;
 
 	private:
-		uint8_t slotTo_;
-		uint8_t slotFrom_;
+		int16_t slotTo_;
+		int16_t slotFrom_;
 };
 
 }
